+++ src/olmo_core/nn/transformer/config.py
@@ -366,7 +366,7 @@ class TransformerConfig(Config):
     def olmo2_370M(cls, vocab_size: int, **kwargs) -> "TransformerConfig":
         return cls.llama_like(
             d_model=1024,
-            hidden_size_multiplier=1.4,
+            hidden_size_multiplier=1.5,
             n_layers=kwargs.pop("n_layers", 16),
             n_heads=kwargs.pop("n_heads", 16),
             vocab_size=vocab_size,
@@ -418,6 +418,7 @@ class TransformerConfig(Config):
             qk_norm=kwargs.pop("qk_norm", True),
             rope_theta=kwargs.pop("rope_theta", 500_000),
             layer_norm_eps=1e-6,
+            hidden_size_multiplier=1.5,
             **kwargs,
         )
 
@@ -425,7 +426,7 @@ class TransformerConfig(Config):
     def olmo2_3B(cls, vocab_size: int, **kwargs) -> "TransformerConfig":
         return cls.llama_like(
             d_model=3328,
-            hidden_size_multiplier=1.4,
+            hidden_size_multiplier=1.5,
             n_layers=kwargs.pop("n_layers", 16),
             n_heads=kwargs.pop("n_heads", 16),
             vocab_size=vocab_size,